requirement.txt
pillow
agno
ollama
google-genai
streamlit